+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/utils/extentions/prayerTime/getHijriReadableDate.kt
@@ -1,25 +1,6 @@
 package net.thechance.mena.faith.presentation.utils.extentions.prayerTime
 
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toLocalDateTime
-import net.thechance.mena.faith.domain.entity.PrayerName
 import net.thechance.mena.faith.domain.entity.PrayerTime
-import kotlin.time.ExperimentalTime
-import kotlin.time.Instant
-
-@OptIn(ExperimentalTime::class)
-fun formatInstantToTimeString(instant: Instant): String {
-    val localDateTime = instant.toLocalDateTime(TimeZone.currentSystemDefault())
-
-    val hour24 = localDateTime.hour
-    val hour12 = if (hour24 % 12 == 0) 12 else hour24 % 12
-    val minute = localDateTime.minute.toString().padStart(2, '0')
-    val amPm = if (hour24 < 12) "AM" else "PM"
-
-    val hourString = hour12.toString().padStart(2, '0')
-    return "$hourString:$minute $amPm"
-}
-
 
 fun getHijriReadableDate(prayerTimes: List<PrayerTime>): String = runCatching {
     val hijriDate = prayerTimes.firstOrNull()?.hijriDate ?: return ""
@@ -32,11 +13,6 @@ fun getHijriReadableDate(prayerTimes: List<PrayerTime>): String = runCatching {
     "$day $monthName $year"
 }.getOrDefault("")
 
-@OptIn(ExperimentalTime::class)
-fun getSunriseTime(prayerTimes: List<PrayerTime>): String =
-    prayerTimes
-        .firstOrNull { it.name == PrayerName.SUNRISE }
-        ?.let { formatInstantToTimeString(it.time) } ?: ""
 
 private val hijriMonths: Map<Int, String> = mapOf(
     1 to "Muharram",
