+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/utils/extentions/prayerTime/getSunriseTime.kt
@@ -1,0 +1,11 @@
+package net.thechance.mena.faith.presentation.utils.extentions.prayerTime
+
+import net.thechance.mena.faith.domain.entity.PrayerName
+import net.thechance.mena.faith.domain.entity.PrayerTime
+import kotlin.time.ExperimentalTime
+
+@OptIn(ExperimentalTime::class)
+fun getSunriseTime(prayerTimes: List<PrayerTime>): String =
+    prayerTimes
+        .firstOrNull { it.name == PrayerName.SUNRISE }
+        ?.let { formatInstantToTimeString(it.time) } ?: ""
